+++ packages/jest-reporters/src/coverage_reporter.ts
@@ -10,8 +10,8 @@
 /// <reference path="./istanbul-api.d.ts" />
 
 import path from 'path';
-import {TestResult, Config} from '@jest/types';
-
+import {Config} from '@jest/types';
+import {AggregatedResult, TestResult} from '@jest/test-result';
 import {clearLine, isInteractive} from 'jest-util';
 import {createReporter} from 'istanbul-api';
 import chalk from 'chalk';
@@ -50,8 +50,8 @@ export default class CoverageReporter extends BaseReporter {
 
   onTestResult(
     _test: Test,
-    testResult: TestResult.TestResult,
-    _aggregatedResults: TestResult.AggregatedResult,
+    testResult: TestResult,
+    _aggregatedResults: AggregatedResult,
   ) {
     if (testResult.coverage) {
       this._coverageMap.merge(testResult.coverage);
@@ -81,7 +81,7 @@ export default class CoverageReporter extends BaseReporter {
 
   async onRunComplete(
     contexts: Set<Context>,
-    aggregatedResults: TestResult.AggregatedResult,
+    aggregatedResults: AggregatedResult,
   ) {
     await this._addUntestedFiles(this._globalConfig, contexts);
     const {map, sourceFinder} = this._sourceMapStore.transformCoverage(