+++ packages/jest-test-result/src/index.ts
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+export {default as formatTestResults} from './formatTestResults';
+export {
+  addResult,
+  buildFailureTestResult,
+  makeEmptyAggregatedTestResult,
+} from './helpers';
+export {
+  AggregatedResult,
+  AssertionLocation,
+  AssertionResult,
+  SerializableError,
+  SnapshotSummary,
+  Status,
+  Suite,
+  TestResult,
+} from './types';